pyproject.toml
[tool.poetry]
name = "pySigma-backend-databricks"
version = "0.1.4"
description = "pySigma Databricks backend"
authors = ["Alex Ott <alexott@gmail.com>"]
license = "MIT"
repository = "https://github.com/alexott/pySigma-backend-databricks"
packages = [
    { include = "sigma" },
    { include = "fieldmapper" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pysigma = "^0.11.23"
inflection = "^0.5.1"
sqlglot = "*"
sigma-cli = "*"
pyyaml = ">=6.0"
openai = ">=1.0.0"
mitreattack-python = {version = ">=3.0.0", python = ">=3.10"}

[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.2"
pylint = "^3.3.8"

[tool.poetry.scripts]
field-mapper = "fieldmapper.ocsf.cli:main"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
disable = "missing-function-docstring,missing-module-docstring,line-too-long,too-many-lines"

[tool.pytest.ini_options]
minversion = "6.0"
python_files = "test_*.py"
addopts = "-vv --cov=sigma --cov-report term --cov-report xml:cov.xml"
testpaths = [
    "tests",
]
